requirements.txt
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
certifi==2024.6.2
cffi==1.16.0
click==8.1.7
coverage==7.3.2
cryptography==42.0.5
distro==1.9.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
et-xmlfile==1.1.0
fastapi==0.110.2
fastapi_auth_middleware==1.0.2
h11==0.14.0
httpcore==1.0.5
httpretty==1.1.4
httpx==0.27.0
idna==3.4
iniconfig==2.0.0
numpy==1.26.4
openai==1.33.0
openpyxl==3.1.2
packaging==23.2
pandas==2.2.2
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.6.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pytest==8.2.2
pytest-cov==4.1.0
pytest-mock==3.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.29
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.21
starlette==0.37.2
tqdm==4.66.4
typing_extensions==4.8.0
tzdata==2024.1
uvicorn==0.23.2
